Move Rotation to data and create utils.typing
Move IndexerType to new mrpro.utils.typing
TYPE_CHECKING guard modify_acq_info to avoid circular imports
Move Rotation to mrpro.data and modify docstring
