+++ src/mrpro/data/__init__.py
@@ -14,5 +14,6 @@ from mrpro.data.KTrajectoryRawShape import KTrajectoryRawShape
 from mrpro.data.MoveDataMixin import MoveDataMixin
 from mrpro.data.QData import QData
 from mrpro.data.QHeader import QHeader
+from mrpro.data.Rotation import Rotation
 from mrpro.data.SpatialDimension import SpatialDimension
 from mrpro.data.TrajectoryDescription import TrajectoryDescription
