+++ src/mrpro/utils/__init__.py
@@ -1,7 +1,7 @@
+import mrpro.utils.slice_profiles
+import mrpro.utils.typing
 from mrpro.utils.smap import smap
 from mrpro.utils.remove_repeat import remove_repeat
 from mrpro.utils.zero_pad_or_crop import zero_pad_or_crop
 from mrpro.utils.modify_acq_info import modify_acq_info
 from mrpro.utils.split_idx import split_idx
-from mrpro.utils.Rotation import Rotation
-import mrpro.utils.slice_profiles
